Implement rest of DRAW instruction
This took me forever -- I could not for the life of me figure out how
to utilize SDL2 correctly. I initially was trying to convert the
display into a texture first via `SDL2::Surface.from_string`, and then
creating a texture from the resulting surface. This went really wrong
in ways I still don't understand.

I based this rendering implementation on this[^1]. I'm disappointed
that I had to look at someone else's implementation to get it working,
but I've been trying to get this to work for so long that I'm just
glad to see the damn IBM logo on my screen.

[^1]: https://github.com/m0tholith/CHIP-8/blob/master/CHIP-8%20Emulator/Renderer.cpp#L60